Add devcontainer
Create a devcontainer to make development and
contribution easier.

The devcontainer is based on the Docker production
image with the exception that it uses puma as
instead of nginx.

I needed to pin some of the gems that were being
used in development as they were causing issues.
Those gems were

byebug
pry-byebug
rubocop